refactor: move File and related functions to scan.file
