+++ ggshield/cmd/secret/scan/path.py
@@ -9,9 +9,9 @@ from ggshield.cmd.secret.scan.secret_scan_common_options import (
 )
 from ggshield.core.constants import MAX_WORKERS
 from ggshield.core.errors import handle_exception
-from ggshield.core.file_utils import get_files_from_paths
 from ggshield.core.text_utils import create_progress_bar
 from ggshield.scan import ScanCollection, ScanContext, ScanMode, SecretScanner
+from ggshield.scan.file import get_files_from_paths
 
 
 @click.command()
