+++ ggshield/cmd/secret/scan/pypi.py
@@ -16,9 +16,9 @@ from ggshield.cmd.secret.scan.secret_scan_common_options import (
 from ggshield.core.config import Config
 from ggshield.core.constants import MAX_WORKERS
 from ggshield.core.errors import UnexpectedError
-from ggshield.core.file_utils import get_files_from_paths
 from ggshield.core.text_utils import create_progress_bar
 from ggshield.scan import Files, ScanCollection, ScanContext, ScanMode, SecretScanner
+from ggshield.scan.file import get_files_from_paths
 
 
 PYPI_DOWNLOAD_TIMEOUT = 30
