+++ ggshield/scan/__init__.py
@@ -1,11 +1,13 @@
 from .commit import Commit
+from .file import File, get_files_from_paths
 from .scan_context import ScanContext
 from .scan_mode import ScanMode
-from .scannable import DecodeError, File, Files, Scannable, StringScannable
+from .scannable import DecodeError, Files, Scannable, StringScannable
 from .scanner import Result, Results, ScanCollection, SecretScanner
 
 
 __all__ = [
+    "get_files_from_paths",
     "Commit",
     "DecodeError",
     "File",
