+++ tests/unit/scan/test_file.py
@@ -4,38 +4,9 @@ from random import randrange
 
 import pytest
 
-from ggshield.core.file_utils import generate_files_from_paths
 from ggshield.scan import DecodeError, File
+from ggshield.scan.file import generate_files_from_paths
 from tests.conftest import is_windows
-
-
-@pytest.mark.parametrize(
-    ["filename", "input_content", "expected_content"],
-    [
-        ("normal.txt", b"Normal", "Normal"),
-        ("invalid-utf8-start-byte.txt", b"Hello\x81World", "Hello·World"),
-        ("zero-bytes-are-kept.txt", b"Zero\0byte", "Zero\0byte"),
-    ],
-)
-def test_generate_files_from_paths(
-    tmp_path, filename: str, input_content: bytes, expected_content: str
-):
-    """
-    GIVEN a file
-    WHEN calling generate_files_from_paths() on it
-    THEN it returns the expected File instance
-    AND the content of the File instance is what is expected
-    """
-    path = tmp_path / filename
-    Path(path).write_bytes(input_content)
-
-    files = list(generate_files_from_paths([str(path)], verbose=False))
-
-    file = files[0]
-    assert file.filename == str(path)
-    assert file.content == expected_content
-
-    assert len(files) == 1
 
 
 @pytest.mark.parametrize(
@@ -180,3 +151,32 @@ def test_file_path():
     str_path = r"c:\Windows" if is_windows() else "/usr"
     file = File(str_path)
     assert file.path == Path(str_path)
+
+
+@pytest.mark.parametrize(
+    ["filename", "input_content", "expected_content"],
+    [
+        ("normal.txt", b"Normal", "Normal"),
+        ("invalid-utf8-start-byte.txt", b"Hello\x81World", "Hello·World"),
+        ("zero-bytes-are-kept.txt", b"Zero\0byte", "Zero\0byte"),
+    ],
+)
+def test_generate_files_from_paths(
+    tmp_path, filename: str, input_content: bytes, expected_content: str
+):
+    """
+    GIVEN a file
+    WHEN calling generate_files_from_paths() on it
+    THEN it returns the expected File instance
+    AND the content of the File instance is what is expected
+    """
+    path = tmp_path / filename
+    Path(path).write_bytes(input_content)
+
+    files = list(generate_files_from_paths([str(path)], verbose=False))
+
+    file = files[0]
+    assert file.filename == str(path)
+    assert file.content == expected_content
+
+    assert len(files) == 1
